+++ src/deploy/functions/prepare.ts
@@ -64,9 +64,8 @@ export async function prepare(
       /* silent=*/ true
     ),
     ensure.cloudBuildEnabled(projectId),
-    ensure.maybeEnableAR(projectId),
+    ensureApiEnabled.ensure(projectId, "artifactregistry.googleapis.com", "artifactregistry"),
   ]);
-  context.artifactRegistryEnabled = checkAPIsEnabled[3];
 
   // Get the Firebase Config, and set it on each function in the deployment.
   const firebaseConfig = await functionsConfig.getFirebaseConfig(options);
@@ -201,7 +200,6 @@ export async function prepare(
     // We'd eventually have to add special error handling for billing APIs, but
     // enableCloudBuild is called above and has this special casing already.
     const V2_APIS = [
-      "artifactregistry.googleapis.com",
       "run.googleapis.com",
       "eventarc.googleapis.com",
       "pubsub.googleapis.com",